+++ Public/ContentProvenanceChecks/Request-ContentProvenanceCheck.ps1
@@ -1,0 +1,103 @@
+function Request-ContentProvenanceCheck {
+    [CmdletBinding()]
+    [OutputType([pscustomobject])]
+    param (
+        [Parameter(Mandatory, Position = 0, ValueFromPipeline)]
+        [ValidateNotNullOrEmpty()]
+        [string]$File,
+
+        [Parameter()]
+        [int]$TimeoutSec = 0,
+
+        [Parameter()]
+        [ValidateRange(0, 100)]
+        [int]$MaxRetryCount = 0,
+
+        [Parameter()]
+        [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
+
+        [Parameter()]
+        [System.Uri]$ApiBase,
+
+        [Parameter(DontShow)]
+        [string]$ApiVersion,
+
+        [Parameter()]
+        [ValidateSet('openai', 'azure', 'azure_ad')]
+        [string]$AuthType = 'openai',
+
+        [Parameter()]
+        [securestring][SecureStringTransformation()]$ApiKey,
+
+        [Parameter()]
+        [Alias('OrgId')]
+        [string]$Organization,
+
+        [Parameter()]
+        [System.Collections.IDictionary]$AdditionalQuery,
+
+        [Parameter()]
+        [System.Collections.IDictionary]$AdditionalHeaders,
+
+        [Parameter()]
+        [object]$AdditionalBody
+    )
+
+    begin {
+        # Get API context
+        $OpenAIParameter = Get-OpenAIAPIParameter -EndpointName 'ContentProvenanceChecks' -Parameters $PSBoundParameters -ErrorAction Stop
+    }
+
+    process {
+        #region Construct parameters for API request
+        $PostBody = [System.Collections.Specialized.OrderedDictionary]::new()
+        $PostBody.file = Resolve-FileInfo $File -ErrorAction Stop
+        #endregion
+
+        #region Send API Request
+        $params = @{
+            Method            = $OpenAIParameter.Method
+            Uri               = $OpenAIParameter.Uri
+            ContentType       = $OpenAIParameter.ContentType
+            TimeoutSec        = $OpenAIParameter.TimeoutSec
+            MaxRetryCount     = $OpenAIParameter.MaxRetryCount
+            ApiKey            = $OpenAIParameter.ApiKey
+            AuthType          = $OpenAIParameter.AuthType
+            Organization      = $OpenAIParameter.Organization
+            Body              = $PostBody
+            AdditionalQuery   = $AdditionalQuery
+            AdditionalHeaders = $AdditionalHeaders
+            AdditionalBody    = $AdditionalBody
+        }
+        $Response = Invoke-OpenAIAPIRequest @params
+
+        # error check
+        if ($null -eq $Response) {
+            return
+        }
+        #endregion
+
+        #region Parse response object
+        try {
+            $Response = $Response | ConvertFrom-Json -ErrorAction Stop
+        }
+        catch {
+            Write-Error -Exception $_.Exception
+            return
+        }
+        #endregion
+
+        #region Output
+        $Response.PSObject.TypeNames.Insert(0, 'PSOpenAI.ContentProvenanceCheck')
+        if ($null -ne $Response.created_at) {
+            # Convert unixtime to [DateTime] for readability, including the Unix epoch.
+            $Response.created_at = [System.DateTimeOffset]::FromUnixTimeSeconds([long]$Response.created_at).LocalDateTime
+        }
+        Write-Output $Response
+        #endregion
+    }
+
+    end {
+
+    }
+}